rename submodules to avoid inception
