+++ rust/xaynet-analytics/src/data_combination/data_points/screen_enter_count.rs
@@ -1,6 +1,6 @@
 use crate::{
-    data_combiner::data_points::data_point::{CalculateDataPoints, DataPointMetadata},
-    repo::analytics_event::{AnalyticsEvent, AnalyticsEventType},
+    data_combination::data_points::data_point::{CalculateDataPoints, DataPointMetadata},
+    data_provision::analytics_event::{AnalyticsEvent, AnalyticsEventType},
 };
 
 // TODO: accept an iterator instead of Vec: https://xainag.atlassian.net/browse/XN-1517
