+++ rust/xaynet-analytics/src/data_combination/data_points/was_active_each_past_period.rs
@@ -2,8 +2,8 @@ use chrono::{DateTime, Utc};
 use std::collections::HashMap;
 
 use crate::{
-    data_combiner::data_points::data_point::{CalculateDataPoints, DataPointMetadata},
-    repo::analytics_event::AnalyticsEvent,
+    data_combination::data_points::data_point::{CalculateDataPoints, DataPointMetadata},
+    data_provision::analytics_event::AnalyticsEvent,
 };
 
 // TODO: accept an iterator instead of Vec: https://xainag.atlassian.net/browse/XN-1517
