+++ rust/xaynet-analytics/src/lib.rs
@@ -2,5 +2,5 @@
 //! a framework that allows mobile applications to collect and aggregate
 //! analytics data via the _Privacy-Enhancing Technology_ (PET) protocol.
 
-pub mod data_combiner;
-pub mod repo;
+pub mod data_combination;
+pub mod data_provision;
